// src/graphql/data/__generated__/types-and-hooks.ts
export enum Chain {
  Arbitrum = "ARBITRUM",
  Avalanche = "AVALANCHE",
  Base = "BASE",
  Bnb = "BNB",
  Celo = "CELO",
  Ethereum = "ETHEREUM",
  EthereumGoerli = "ETHEREUM_GOERLI",
  EthereumSepolia = "ETHEREUM_SEPOLIA",
  Optimism = "OPTIMISM",
  Polygon = "POLYGON",
  UnknownChain = "UNKNOWN_CHAIN"
}
// ...
export enum TokenStandard {
  Erc20 = "ERC20",
  Native = "NATIVE"
}